Fix dtypes causing overflow issues with accumulators

+++ src/notscared/distinguishers/cpa.py
@@ -1,6 +1,6 @@
 import numpy as np
 import math as math
-from src.notscared.utils.leakage import create_leakage_table
+from utils.leakage import create_leakage_table
 # from statistics.welford import Welford
 
 class CPA:
@@ -26,11 +26,11 @@ class CPA:
             leakage_cube[trace_index] = create_leakage_table(plaintext[trace_index], self.use_hamming_weight)
         for byte in range(self.byte_range[0], self.byte_range[1]):
             for leakage_key_hypothesis in range(leakage_cube.shape[2]):
-                self.leakage_accs[byte, leakage_key_hypothesis, 0] += np.sum(leakage_cube[:, byte, leakage_key_hypothesis])
-                self.leakage_accs[byte, leakage_key_hypothesis, 1] += np.sum(np.square(leakage_cube[:, byte, leakage_key_hypothesis], dtype=np.float32))
+                self.leakage_accs[byte, leakage_key_hypothesis, 0] += np.sum(leakage_cube[:, byte, leakage_key_hypothesis], dtype=np.float32)
+                self.leakage_accs[byte, leakage_key_hypothesis, 1] += np.sum(np.square(leakage_cube[:, byte, leakage_key_hypothesis], dtype=np.float32), dtype=np.float32)
         for point_in_time in range(traces.shape[1]):
-            self.trace_accs[point_in_time, 0] += np.sum(traces[:, point_in_time])
-            self.trace_accs[point_in_time, 1] += np.sum(np.square(traces[:, point_in_time]))
+            self.trace_accs[point_in_time, 0] += np.sum(traces[:, point_in_time], dtype=np.float32)
+            self.trace_accs[point_in_time, 1] += np.sum(np.square(traces[:, point_in_time], dtype=np.float32), dtype=np.float32)
             for byte in range(self.byte_range[0], self.byte_range[1]):
                 for leakage_key_hypothesis in range(leakage_cube.shape[2]):
                     trace_slice = traces[:, point_in_time]
